普通用户注册

+++ piesat-school-biz/src/main/java/com/piesat/school/biz/ds/user/mapper/RoleMapper.java
@@ -12,5 +12,5 @@ import java.util.Set;
  */
 
 public interface RoleMapper extends BaseMapper<Role> {
-    Set<Role> getRolesByUserId(Integer id);
+    Set<Role> getRolesByUserId(Long id);
 }
